#6 Fix Event entity due DB table

+++ restservice/src/main/java/org/itevents/model/Event.java
@@ -12,7 +12,7 @@ public class Event implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @XmlElement
-    private long id;
+    private int id;
     @XmlElement
     private String title;
     @XmlElement
@@ -31,7 +31,7 @@ public class Event implements Serializable {
     public Event() {
     }
 
-    public Event(long id, String title, Date eventDate, Date createDate, String regLink, String address, Location location,  String contact) {
+    public Event(int id, String title, Date eventDate, Date createDate, String regLink, String address, Location location,  String contact) {
         this.id = id;
         this.title = title;
         this.eventDate = eventDate;
@@ -42,11 +42,11 @@ public class Event implements Serializable {
         this.contact = contact;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
